backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
python-multipart
aiosqlite
google-genai
ultralytics
python-dotenv
